refactor(core): split installer boundaries
